refactor(admin): isolate offline CLI composition
